Added manual image verification applet.
Intention of the this applet is so that users can perform a final image
verification step to remove any bad results, exempli gratia blurry
images.

Information about the camera is now written to a json file.
path is now configurable, this is to accommodate future user
configuration.

Added progress information in the recording step.